refactor(MetricScene): Extract MetricActionBar
