--- src/ActionTabs/RelatedMetricsScene.tsx
@@ -1,5 +1,0 @@
-import { MetricSelectScene } from '../MetricSelect/MetricSelectScene';
-
-export function buildRelatedMetricsScene() {
-  return new MetricSelectScene({});
-}
