+++ src/MetricGraphScene.tsx
@@ -13,8 +13,9 @@ import {
 import { useStyles2 } from '@grafana/ui';
 import React from 'react';
 
+import { MetricActionBar } from 'MetricActionBar';
+
 import { AutoVizPanel } from './autoQuery/components/AutoVizPanel';
-import { MetricActionBar } from './MetricScene';
 import { getTrailFor, getTrailSettings } from './utils';
 
 export const MAIN_PANEL_MIN_HEIGHT = 280;
