+++ src/MetricScene.tsx
@@ -1,10 +1,6 @@
-import { css } from '@emotion/css';
-import { type GrafanaTheme2 } from '@grafana/data';
 import { config } from '@grafana/runtime';
 import {
-  getExploreURL,
   QueryVariable,
-  sceneGraph,
   SceneObjectBase,
   SceneObjectUrlSyncConfig,
   SceneVariableSet,
@@ -14,27 +10,17 @@ import {
   type SceneObjectState,
   type SceneObjectUrlValues,
 } from '@grafana/scenes';
-import { Box, Icon, LinkButton, Stack, Tab, TabsBar, ToolbarButton, Tooltip, useStyles2 } from '@grafana/ui';
 import React from 'react';
 
-import { buildRelatedMetricsScene } from './ActionTabs/RelatedMetricsScene';
-import { AutoVizPanel } from './autoQuery/components/AutoVizPanel';
+import { actionViews, actionViewsDefinitions, type ActionViewType } from 'MetricActionBar';
+
 import { getAutoQueriesForMetric } from './autoQuery/getAutoQueriesForMetric';
 import { type AutoQueryDef, type AutoQueryInfo } from './autoQuery/types';
-import { buildLabelBreakdownActionScene } from './Breakdown/LabelBreakdownScene';
-import { UI_TEXT } from './constants/ui';
-import { reportExploreMetrics } from './interactions';
-import {
-  MAIN_PANEL_MAX_HEIGHT,
-  MAIN_PANEL_MIN_HEIGHT,
-  METRIC_AUTOVIZPANEL_KEY,
-  MetricGraphScene,
-} from './MetricGraphScene';
+import { MAIN_PANEL_MAX_HEIGHT, MAIN_PANEL_MIN_HEIGHT, MetricGraphScene } from './MetricGraphScene';
 import { RelatedLogsOrchestrator } from './RelatedLogs/RelatedLogsOrchestrator';
 import { buildRelatedLogsScene } from './RelatedLogs/RelatedLogsScene';
 import {
   getVariablesWithMetricConstant,
-  MetricSelectedEvent,
   RefreshMetricsEvent,
   trailDS,
   VAR_FILTERS,
@@ -42,9 +28,6 @@ import {
   VAR_METRIC_EXPR,
   type MakeOptional,
 } from './shared';
-import { ShareTrailButton } from './ShareTrailButton';
-import { useBookmarkState } from './TrailStore/useBookmarkState';
-import { getTrailFor, getUrlForTrail } from './utils';
 
 export interface MetricSceneState extends SceneObjectState {
   body: MetricGraphScene;
@@ -55,14 +38,6 @@ export interface MetricSceneState extends SceneObjectState {
   queryDef?: AutoQueryDef;
   relatedLogsCount?: number;
 }
-
-export const actionViews = {
-  breakdown: 'breakdown',
-  related: 'related',
-  relatedLogs: 'logs',
-} as const;
-
-export type ActionViewType = (typeof actionViews)[keyof typeof actionViews];
 
 export class MetricScene extends SceneObjectBase<MetricSceneState> {
   public readonly relatedLogsOrchestrator = new RelatedLogsOrchestrator(this);
@@ -158,168 +133,6 @@ export class MetricScene extends SceneObjectBase<MetricSceneState> {
       orchestrator: this.relatedLogsOrchestrator,
     });
   }
-}
-
-interface ActionViewDefinition {
-  displayName: string;
-  value: ActionViewType;
-  description?: string;
-  getScene: (metricScene: MetricScene) => SceneObject<SceneObjectState>;
-}
-
-const actionViewsDefinitions: ActionViewDefinition[] = [
-  { displayName: 'Breakdown', value: actionViews.breakdown, getScene: buildLabelBreakdownActionScene },
-  {
-    displayName: 'Related metrics',
-    value: actionViews.related,
-    getScene: buildRelatedMetricsScene,
-    description: 'Relevant metrics based on current label filters',
-  },
-  {
-    displayName: 'Related logs',
-    value: actionViews.relatedLogs,
-    getScene: (metricScene: MetricScene) => metricScene.createRelatedLogsScene(),
-    description: 'Relevant logs based on current label filters and time range',
-  },
-];
-
-export interface MetricActionBarState extends SceneObjectState {}
-
-export class MetricActionBar extends SceneObjectBase<MetricActionBarState> {
-  public getLinkToExplore = async () => {
-    const metricScene = sceneGraph.getAncestor(this, MetricScene);
-    const autoVizPanel = sceneGraph.findByKeyAndType(this, METRIC_AUTOVIZPANEL_KEY, AutoVizPanel);
-    const panelData =
-      typeof autoVizPanel.state.panel !== 'undefined'
-        ? sceneGraph.getData(autoVizPanel.state.panel).state.data
-        : undefined;
-
-    if (!panelData) {
-      throw new Error('Cannot get link to explore, no panel data found');
-    }
-
-    return getExploreURL(panelData, metricScene, panelData.timeRange);
-  };
-
-  public openExploreLink = async () => {
-    reportExploreMetrics('selected_metric_action_clicked', { action: 'open_in_explore' });
-    this.getLinkToExplore().then((link) => {
-      // We use window.open instead of a Link or <a> because we want to compute the explore link when clicking,
-      // if we precompute it we have to keep track of a lot of dependencies
-      window.open(link, '_blank');
-    });
-  };
-
-  public static readonly Component = ({ model }: SceneComponentProps<MetricActionBar>) => {
-    const metricScene = sceneGraph.getAncestor(model, MetricScene);
-    const styles = useStyles2(getStyles);
-    const trail = getTrailFor(model);
-    const [isBookmarked, toggleBookmark] = useBookmarkState(trail);
-    const { actionView } = metricScene.useState();
-
-    return (
-      <Box paddingY={1} data-testid="action-bar">
-        <div className={styles.actions}>
-          <Stack gap={1}>
-            <ToolbarButton
-              variant={'canvas'}
-              tooltip={UI_TEXT.METRIC_SELECT_SCENE.SELECT_NEW_METRIC_TOOLTIP}
-              onClick={() => {
-                reportExploreMetrics('selected_metric_action_clicked', { action: 'unselect' });
-                trail.publishEvent(new MetricSelectedEvent(undefined));
-              }}
-            >
-              Select new metric
-            </ToolbarButton>
-            <ToolbarButton
-              variant={'canvas'}
-              icon="compass"
-              tooltip={UI_TEXT.METRIC_SELECT_SCENE.OPEN_EXPLORE_LABEL}
-              onClick={model.openExploreLink}
-            />
-            <ShareTrailButton trail={trail} />
-            <ToolbarButton
-              variant={'canvas'}
-              icon={
-                isBookmarked ? (
-                  <Icon name={'favorite'} type={'mono'} size={'lg'} />
-                ) : (
-                  <Icon name={'star'} type={'default'} size={'lg'} />
-                )
-              }
-              tooltip={UI_TEXT.METRIC_SELECT_SCENE.BOOKMARK_LABEL}
-              onClick={toggleBookmark}
-            />
-            {trail.state.embedded && (
-              <LinkButton
-                href={getUrlForTrail(trail)}
-                variant={'secondary'}
-                onClick={() => reportExploreMetrics('selected_metric_action_clicked', { action: 'open_from_embedded' })}
-              >
-                Open
-              </LinkButton>
-            )}
-          </Stack>
-        </div>
-
-        <TabsBar className={styles.customTabsBar}>
-          {actionViewsDefinitions.map((tab, index) => {
-            const label = tab.displayName;
-            const counter = tab.value === actionViews.relatedLogs ? metricScene.state.relatedLogsCount : undefined;
-            const isActive = actionView === tab.value;
-
-            const tabRender = (
-              <Tab
-                key={index}
-                label={label}
-                counter={counter}
-                active={isActive}
-                onChangeTab={() => {
-                  if (isActive) {
-                    return;
-                  }
-
-                  reportExploreMetrics('metric_action_view_changed', {
-                    view: tab.value,
-                    related_logs_count: metricScene.relatedLogsOrchestrator.checkConditionsMetForRelatedLogs()
-                      ? counter
-                      : undefined,
-                  });
-
-                  metricScene.setActionView(tab.value);
-                }}
-              />
-            );
-
-            if (tab.description) {
-              return (
-                <Tooltip key={index} content={tab.description} placement="bottom-start" theme="info">
-                  {tabRender}
-                </Tooltip>
-              );
-            }
-            return tabRender;
-          })}
-        </TabsBar>
-      </Box>
-    );
-  };
-}
-
-function getStyles(theme: GrafanaTheme2) {
-  return {
-    actions: css({
-      [theme.breakpoints.up(theme.breakpoints.values.md)]: {
-        position: 'absolute',
-        right: 0,
-        top: 16,
-        zIndex: 2,
-      },
-    }),
-    customTabsBar: css({
-      paddingBottom: theme.spacing(1),
-    }),
-  };
 }
 
 function getVariableSet(metric: string) {
